сделела исправления
