+++ src/components/Column/Column.jsx
@@ -1,4 +1,4 @@
-import Card from "../Card/Card";
+import Card from "./Card/Card";
 
 function Column ({title}){
   return(
